pyproject.toml
[project]
name = "coproxy-ai"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "cryptography>=42.0",
]

[project.optional-dependencies]
keyring = ["keyring>=25.0"]
yaml = ["pyyaml>=6.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff",
]

[project.scripts]
coproxy = "coproxy.main:cli"
coproxy-login = "coproxy.auth.login:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
